@@ -1,18 +1,18 @@
 name: robotremoteserver_dart
 description: Robot framework remote server
-version: 0.0.2
+version: 0.0.3
 homepage: https://github.com/binithb/robotremoteserver_dart
 
 environment:
-  sdk: '>=2.10.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  collection: ^1.15.0
-  http: ^0.12.1
-  meta: ^1.3.0
-  xml: ^4.1.0
-  xml_rpc: ^0.3.0
+  collection:
+  http:
+  meta:
+  xml:
+  xml_rpc: 0.4.0
 
 dev_dependencies:
-  pedantic: ^1.9.0
-  test: ^1.14.4
+  pedantic:
+  test:
