@@ -14,38 +14,40 @@ final robotCodecs = List<Codec>.unmodifiable(<Codec>[
   mapCodec,
 ]);
 
-typedef XmlCodecDecodeSignature = Object Function(XmlNode);
-typedef XmlCodecEncodeSignature = XmlNode Function(Object);
+typedef XmlCodecDecodeSignature = Object? Function(XmlNode);
+typedef XmlCodecEncodeSignature = XmlNode Function(Object?);
 
 final mapCodec = _MapCodec();
 
 class _MapCodec implements Codec<Map<String, dynamic>> {
   @override
-  XmlNode encode(value, XmlCodecEncodeSignature encode) {
+  XmlNode encode(Object? value, XmlCodecEncodeSignature? encode) {
     // if (value is Map<dynamic, dynamic>) throw ArgumentError();
+    if (value is! Map<dynamic, dynamic?>) throw ArgumentError();
 
     final members = <XmlNode>[];
     value.forEach((k, v) {
       members.add(XmlElement(XmlName('member'), [], [
         XmlElement(XmlName('name'), [], [XmlText(k)]),
-        XmlElement(XmlName('value'), [], [encode(v)])
+        XmlElement(XmlName('value'), [], [encode!(v)])
       ]));
     });
     return XmlElement(XmlName('struct'), [], members);
   }
 
   @override
-  Map<String, dynamic> decode(XmlNode node, XmlCodecDecodeSignature decode) {
-    if ((node is XmlElement && node.name.local == 'struct')) {
+  Map<String, dynamic> decode(XmlNode? node, XmlCodecDecodeSignature? decode) {
+    if (!(node is XmlElement && node.name.local == 'struct')) {
       throw ArgumentError();
     }
-
     final struct = <String, dynamic>{};
     for (final member in node.findElements('member')) {
       final name = member.findElements('name').first.text;
       final valueElt = member.findElements('value').first;
       final elt = getValueContent(valueElt);
-      struct[name] = decode(elt);
+      if (null != elt) {
+        struct[name] = decode!(elt);
+      }
     }
     return struct;
   }
